move elements that were a single file in folders to their own files
