+++ client/js/elements/add-files.js
