+++ client/js/elements/gravatar.js
