+++ client/js/elements/loader-icon.js
